#17 completed book summary page

+++ app/src/main/java/com/example/bookshelfapp/data/features/books/repository/remote/model/BooksItem.kt
@@ -15,7 +15,7 @@ data class BooksItem(
     @SerializedName("image")
     val image: String?,
     @SerializedName("lastChapterDate")
-    val lastChapterDate: Int?,
+    val lastChapterDate: Long?,
     @SerializedName("title")
     val title: String?,
     var isFav: Boolean?
